Total saving of about 46 less minutes spent building. 

Travis was taking 16 min 4 sec (total time: 1 hour 23) (not including mac builds). Now it takes 23 min 53 sec (total time: 43 minutes 52 sec). It used to have 10 jobs, now it has 4 jobs.

Appveyor was taking 21 minutes 51 seconds. Now it takes 15 minutes 43.


- Turn on mac builds for PRs on travis (using precompiled conan deps that are not suitable for release, but fine for running tests)
- Compile in parallel with -j4 (on travis linux, appveyor and manylinux).
- Make python2 setup.py not crash with a -j4 flag.
- Exclude slow and unreliable timing tests on linux, appveyor and manylinux builds. Closes https://github.com/pygame/pygame/issues/2109 VMs have unexpected time slices.
- Only include SDL1 dependencies on linux SDL1 builds, and do not install SDL1 deps on SDL2 builds.
- On manylinux job do not run linux tests, install apt deps, or test generating an sdist.
- Only test creating an sdist once on travis (and nowhere else).
- Removed some of the linux python builds on travis, because we cover all the pythons with the manylinux job on there. The trade off is faster builds, for a slightly more confusing situation finding the error. Probably appveyor will also fail for python related issues, and that has individual python jobs.
- Commented out the s390x, and ppc64le builds. We can enable them closer to releases, or whenever people are doing C code with possible endian issues.
- There are 4 total Travis CI builds now. With 5 concurrent jobs, one pull request will take 4 jobs leaving 1 job for another PR. Also only 2 of the jobs are slow ones. This means 3 PRs all started at once can have at least 1 job finished within 8 minutes.
- New version of cibuildwheel on mac, now python 3.9 builds are actually done.